fixed typo and moved function to other file
